Added token generation compatibility for Linux & MacOS
Linux and MacOS use UTF-8 rather than Windows Powershell UTF-16.  Due to the extra null bytes from UTF-16, password conversion from SecureString to plaintext was terminating after the first character on Linux.  This commit adds a check to determine if the host system is Windows or Linux/MacOS.  If Windows, the original logic runs.  If Linux or MacOS, a different method for converting a SecureString to plaintext is performed.